backend/pyproject.toml
[tool.poetry]
name = "academiq-backend"
version = "0.1.0"
description = "AcademIQ Reviewer FastAPI backend"
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
python-dotenv = "^1.0.0"
httpx = "^0.27.0"
PyPDF2 = "^3.0.0"
openai = "^1.40.0"  # if using OpenAI later
supabase = "^2.4.0" # supabase-py client
pydantic = "^2.7.0"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.0"
ruff = "^0.5.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
